Updated logic for exercise times as well as what is required to meet goals. Also basic xaml updated to reflect this with bindings.

+++ Simple Exercise Tracker/Simple Exercise Tracker/Models/ExerciseLog.cs
@@ -8,7 +8,7 @@ namespace Simple_Exercise_Tracker.Models
     public class ExerciseLog
     {
         public DateTime Date { get; set; }
-        public int MinutesExercised { get; set; }
+        public double MinutesExercised { get; set; }
 
     }
 }
